requirements.txt
grobid-client-python
tqdm
requests
